Fixed profile image not showing and moved multer to separate file
