+++ client/src/components/CreateClass/ImageUpload.scss
@@ -1,8 +1,8 @@
-input[type='file'] {
+.custom-file-upload {
   display: none;
 }
 
-.custom-file-upload {
+.custom-file-upload-label {
   border: 1px solid #ccc;
   display: inline-block;
   padding: 6px 12px;
