+++ client/src/components/Profile/styles.scss
@@ -44,4 +44,3 @@
 .profile .form {
   margin-bottom: 30px;
 }
-
